WIP - new Dispatcher
In preparation for 1.0.0 a Dispatcher object has been created which
takes over a fair amount of work that the API was previously. The goal
is to reduce the complexity and object lifespan where possible to keep
things tight and performant while at the same time giving us the best
possible chance at maintenance.

Still very much WIP.